+++ changelog.md
@@ -1,4 +1,8 @@
 ## changelog for elevator-system-simulation-frontend
 
+## 0.0.2
+- Create connection to backend
+- Display building info and elevators
+
 ## 0.0.1
 - Init project